+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 * Replace DraftJS Toolbar plugin H1/H2 buttons for H2/H3 ones @sneridagh
 * Sync i18n translations @sneridagh
 * Fix CSS .input class scope intrusion on the project introduced by the AnchorLink plugin fork @sneridagh
+* Improve search reducer by adding the batching property in the search store.
 
 ## 0.5.0 (2018-03-23)